Slightly modify tweet scraping to enable resumption

+++ scripts/scraping_tweets.py
@@ -3,7 +3,7 @@
 import snscrape.modules.twitter as sntwitter
 #import csv
 import pandas as pd
-from random import randint
+from random import uniform
 from time import sleep
 import pathlib
 
@@ -25,7 +25,14 @@ import pathlib
 
 
 
-def get_user_tweets(username, start_date, end_date, max_tweets, outdir, **kwargs):
+def get_user_tweets(
+    username, 
+    start_date, 
+    end_date, 
+    max_tweets, 
+    outdir, 
+    sleep_sec, 
+    **kwargs):
 
 	'''
 	Function that uses the snscrape module to scrape tweets
@@ -62,74 +69,83 @@ def get_user_tweets(username, start_date, end_date, max_tweets, outdir, **kwargs
 
 	username = username.replace('@', '')
 
-    print("Scraping tweets by user", username)
-
-	query = f'from:@{username} + since:{start_date} until:{end_date}-filter:links -filter:replies'
-
-	tweet_generator = sntwitter.TwitterSearchScraper(query).get_items()
-
-    tweet_dict = {"username" : [], "date" : [], "tweet_content" : []}
-
-    for tweet in tweet_generator:
-
-        tweet_dict["username"].append(tweet.user.username)
-        tweet_dict["date"].append(tweet.date)
-        tweet_dict["tweet_content"].append(tweet.content)
-
-    tweet_df = pd.DataFrame(tweet_dict)
-
-    for col, val in kwargs.items():
-
-        tweet_df[col] = val
-
-    ## 
-
     daterange_string = start_date + "_" + end_date
-
     filename = username + "_" + daterange_string + ".csv"
-
     filepath = pathlib.Path(outdir).joinpath(daterange_string, filename)
-
     filepath.parent.mkdir(parents=True, exist_ok=True)
 
-    tweet_df.to_csv(filepath)
+    if not filepath.is_file():
 
-    print("Scraped", tweet_df.shape[0], "tweets by user", username)
+        print("Scraping tweets by user", username)
+
+    	query = f'from:@{username} + since:{start_date} until:{end_date}-filter:links -filter:replies'
+
+    	tweet_generator = sntwitter.TwitterSearchScraper(query).get_items()
+
+        tweet_dict = {"username" : [], "date" : [], "tweet_content" : []}
+
+        for tweet in tweet_generator:
+
+            tweet_dict["username"].append(tweet.user.username)
+            tweet_dict["date"].append(tweet.date)
+            tweet_dict["tweet_content"].append(tweet.content)
+
+        tweet_df = pd.DataFrame(tweet_dict)
+
+        for col, val in kwargs.items():
+
+            tweet_df[col] = val
+
+        ## 
+
+        tweet_df.to_csv(filepath)
+
+        print("Scraped", tweet_df.shape[0], "tweets by user", username)
+
+        sleep(uniform(sleep_sec, sleep_sec + 5))
+
+    else:
+
+        print("Tweets by user", username, "already scraped.  Reading file . . .")
+
+        tweet_df = pd.read_csv(filepath)
+
+        print("Read file containing", tweet_df.shape[0], "tweets by user", username)
 
     return tweet_df
 
 
-def get_multi_user_tweets(usernames, start_date, end_date, max_tweets, outdir, sleep_sec = 10):
+def get_multi_user_tweets(
+    usernames, start_date, end_date, max_tweets, outdir, sleep_sec, **kwargs):
 
     dfs = [None]*len(usernames)
 
     for i, user in enumerate(usernames):
 
-        dfs[i] = get_user_tweets(user, start_date, end_date, max_tweets, outdir)
-
-        sleep(randint(sleep_sec, sleep_sec + 30))
+        dfs[i] = get_user_tweets(
+            user, start_date, end_date, max_tweets, outdir, sleep_sec, **kwargs)
 
     concat_df = pd.concat(dfs, axis = 1)
 
     return concat_df
 
-def get_democrat_tweets(handles_df, start_date, end_date, max_tweets, outdir, sleep_sec):
+def get_democrat_tweets(
+    handles_df, start_date, end_date, max_tweets, outdir, sleep_sec):
 
     dem_handles = handles_df.Twitter[handles_df.Party.isin(["D", "I"])]
 
-    dem_tweets = get_multi_user_tweets(dem_handles, start_date, end_date, max_tweets, outdir, sleep_sec)
-
-    dem_tweets["party"] = "D"
+    dem_tweets = get_multi_user_tweets(
+        dem_handles, start_date, end_date, max_tweets, outdir, sleep_sec, party = "D")
 
     return dem_tweets
 
-def get_republican_tweets(handles_df, start_date, end_date, max_tweets, outdir, sleep_sec):
+def get_republican_tweets(
+    handles_df, start_date, end_date, max_tweets, outdir, sleep_sec):
 
     rep_handles = handles_df.Twitter[handles_df.Party.eq("R")]
 
-    rep_tweets = get_multi_user_tweets(rep_handles, start_date, end_date, max_tweets, outdir, sleep_sec)
-
-    rep_tweets["party"] = "R"
+    rep_tweets = get_multi_user_tweets(
+        rep_handles, start_date, end_date, max_tweets, outdir, sleep_sec, party = "R")
 
     return rep_tweets  
 
@@ -147,7 +163,7 @@ def get_all_tweets(handles_df, start_date, end_date, max_tweets, outdir, sleep_s
 handles_df = pd.read_csv("/home/aw/projects/congress_twitter/data/congress_twitter_handles.csv")
 
 
-all_tweets = get_all_tweets(handles_df, "2020-01-01", "2020-12-31", max_tweets = 5000, outdir = "./output/", sleep_sec = 30)
+all_tweets = get_all_tweets(handles_df, "2020-01-01", "2020-12-31", max_tweets = 5000, outdir = "./output/", sleep_sec = 5)
